UnitTest update to increase test coverage

+++ Frends.FTP.UploadFiles/Frends.FTP.UploadFiles.Tests/UploadFilesTests.cs
@@ -13,7 +13,7 @@ namespace Frends.FTP.UploadFiles.Tests
     {
         private string _dir;
 
-        private readonly Source _source = new()
+        private Source _source = new()
         {
             Directory = default,
             FileName = default,
@@ -24,32 +24,50 @@ namespace Frends.FTP.UploadFiles.Tests
             Operation = default
         };
 
-        private readonly Connection _connection = new()
+        private Connection _connection = new()
         {
-            Address = default,
-            UserName = default,
-            Password = default,
-            Port = default,
-            EnableClientAuth = default,
-            KeepConnectionAliveInterval = default,
-            ValidateAnyCertificate = default,
-            BufferSize = default,
-            CertificateHashStringSHA1 = default,
-            ClientCertificatePath = default,
-            ConnectionTimeout = default,
-            Encoding = default,
-            Mode = default,
-            SecureDataChannel = default,
-            SslMode = default,
+            Address = Helpers.FtpHost,
+            UserName = Helpers.FtpUsername,
+            Password = Helpers.FtpPassword,
+            Port = Helpers.FtpPort,
             TransportType = default,
-            UseFTPS = default,
+            Mode = default,
+            KeepConnectionAliveInterval = default,
+            ConnectionTimeout = 60,
+            Encoding = default,
+            BufferSize = default,
+            UseFTPS = false,
+            SslMode = default,
+            CertificateHashStringSHA1 = default,
+            EnableClientAuth = default,
+            ValidateAnyCertificate = default,
+            ClientCertificatePath = default,
+            SecureDataChannel = false,
         };
 
-        private readonly Destination _destination = new()
+        private Destination _destination = new()
         {
             Directory = default,
             Action = default,
             FileName = default,
+        };
+
+        private Options _options = new()
+        {
+            CreateDestinationDirectories = default,
+            OperationLog = default,
+            PreserveLastModified = default,
+            RenameDestinationFileDuringTransfer = default,
+            RenameSourceFileBeforeTransfer = default,
+            ThrowErrorOnFail = default
+        };
+
+        private Info _info = new()
+        {
+            ProcessUri = default,
+            TaskExecutionID = default,
+            TransferName = default,
+            WorkDir = default,
         };
 
         private void CreateDummyFileInDummyDir(string fileName)
@@ -67,7 +85,8 @@ namespace Frends.FTP.UploadFiles.Tests
         [TearDown]
         public void TearDown()
         {
-            Directory.Delete(_dir, true);
+            if (Directory.Exists(_dir))
+                Directory.Delete(_dir, true);
         }
 
         [Test]
@@ -77,26 +96,20 @@ namespace Frends.FTP.UploadFiles.Tests
 
             foreach (var destinationAction in destinationActions)
             {
-                // Setup
                 SetUp();
-                CreateDummyFileInDummyDir("file1.txt");
+                var fileName = @$"file{Guid.NewGuid()}.txt";
+                CreateDummyFileInDummyDir(fileName);
+
                 var source = _source;
                 source.Directory = _dir;
-                source.FileName = "file1.txt";
+                source.FileName = fileName;
 
                 var destination = _destination;
                 destination.Directory = "/";
                 destination.Action = destinationAction;
-                destination.FileName = "fileName";
-
-                var connection = _connection;
-                connection.Address = Helpers.FtpHost;
-                connection.UserName = Helpers.FtpUsername;
-                connection.Password = Helpers.FtpPassword;
-                connection.Port = Helpers.FtpPort;
 
                 // Test and assert
-                var result = FTP.UploadFiles(source, destination, connection, new Options(), new Info(), new CancellationToken());
+                var result = FTP.UploadFiles(source, destination, _connection, _options, _info, new CancellationToken());
                 Assert.IsTrue(result.Success);
                 Assert.AreEqual(1, result.SuccessfulTransferCount);
                 TearDown();
@@ -107,37 +120,34 @@ namespace Frends.FTP.UploadFiles.Tests
         public void UploadFTPS_CorrectFingerprint()
         {
             var destinationActions = new[] { DestinationAction.Error, DestinationAction.Overwrite, DestinationAction.Append };
-            var ftpsSslModes = new[] { FtpsSslMode.None, FtpsSslMode.Explicit, FtpsSslMode.Explicit, FtpsSslMode.None, FtpsSslMode.Auto };
+            var ftpsSslModes = new[] { FtpsSslMode.None, FtpsSslMode.Explicit, FtpsSslMode.Auto };
             foreach (var destinationAction in destinationActions)
             {
-                SetUp();
-                // Setup
-                CreateDummyFileInDummyDir("file1.txt");
-                var source = _source;
-                source.Directory = _dir;
-                source.FileName = "file1.txt";
-
-
-                var destination = _destination;
-                destination.Directory = "/";
-                destination.Action = DestinationAction.Overwrite;
 
                 foreach (var ftpsSslMode in ftpsSslModes)
                 {
+                    SetUp();
+                    var fileName = @$"file{Guid.NewGuid()}.txt";
+                    CreateDummyFileInDummyDir(fileName);
+
+                    var source = _source;
+                    source.Directory = _dir;
+                    source.FileName = fileName;
+
+                    var destination = _destination;
+                    destination.Directory = "/";
+                    destination.Action = destinationAction;
+
                     // Our test certificate hashes:
                     // SHA-256: 90:bc:7f:71:14:f5:c2:ad:03:46:d6:ff:75:d5:fe:12:ba:74:23:73:54:31:70:60:b4:8b:bd:8e:87:21:9c:16
                     // SHA-1: d9:11:26:29:84:de:9c:c3:2a:35:18:a1:09:4c:d2:42:49:ea:5c:49
                     var connection = _connection;
-                    connection.Address = Helpers.FtpHost;
-                    connection.UserName = Helpers.FtpUsername;
-                    connection.Password = Helpers.FtpPassword;
-                    connection.Port = Helpers.FtpPort;
                     connection.SslMode = ftpsSslMode;
                     connection.UseFTPS = true;
                     connection.CertificateHashStringSHA1 = "D911262984DE9CC32A3518A1094CD24249EA5C49";
 
                     // Test and assert
-                    var result = FTP.UploadFiles(source, destination, connection, new Options(), new Info(), new CancellationToken());
+                    var result = FTP.UploadFiles(source, destination, connection, _options, _info, new CancellationToken());
                     Assert.IsTrue(result.Success);
                     Assert.AreEqual(1, result.SuccessfulTransferCount);
                     TearDown();
@@ -148,47 +158,33 @@ namespace Frends.FTP.UploadFiles.Tests
         [Test]
         public void UploadFTPS_IncorrectFingerprint()
         {
-            var destinationActions = new[] { DestinationAction.Error, DestinationAction.Overwrite, DestinationAction.Append };
-            var ftpsSslModes = new[] { FtpsSslMode.None, FtpsSslMode.Explicit, FtpsSslMode.Explicit, FtpsSslMode.None, FtpsSslMode.Auto };
-            foreach (var destinationAction in destinationActions)
+            var fileName = @$"file{Guid.NewGuid()}.txt";
+            CreateDummyFileInDummyDir(fileName);
+
+            var source = _source;
+            source.Directory = _dir;
+            source.FileName = fileName;
+
+            var destination = _destination;
+            destination.Directory = "/";
+
+            // Our test certificate hashes:
+            // SHA-256: 90:bc:7f:71:14:f5:c2:ad:03:46:d6:ff:75:d5:fe:12:ba:74:23:73:54:31:70:60:b4:8b:bd:8e:87:21:9c:16
+            // SHA-1: d9:11:26:29:84:de:9c:c3:2a:35:18:a1:09:4c:d2:42:49:ea:5c:49
+            var connection = _connection;
+            connection.SslMode = FtpsSslMode.Explicit;
+            connection.UseFTPS = true;
+            connection.CertificateHashStringSHA1 = "incorrect";
+
+            // Test and assert
+            var ex = Assert.Throws<AggregateException>(() =>
             {
-                SetUp();
-                // Setup
-                CreateDummyFileInDummyDir("file1.txt");
-                var source = _source;
-                source.Directory = _dir;
-                source.FileName = "file1.txt";
+                var result = FTP.UploadFiles(source, destination, connection, new Options(), new Info(),
+                    new CancellationToken());
+            });
 
-
-                var destination = _destination;
-                destination.Directory = "/";
-                destination.Action = DestinationAction.Overwrite;
-
-                foreach (var ftpsSslMode in ftpsSslModes)
-                {
-                    // Our test certificate hashes:
-                    // SHA-256: 90:bc:7f:71:14:f5:c2:ad:03:46:d6:ff:75:d5:fe:12:ba:74:23:73:54:31:70:60:b4:8b:bd:8e:87:21:9c:16
-                    // SHA-1: d9:11:26:29:84:de:9c:c3:2a:35:18:a1:09:4c:d2:42:49:ea:5c:49
-                    var connection = _connection;
-                    connection.Address = Helpers.FtpHost;
-                    connection.UserName = Helpers.FtpUsername;
-                    connection.Password = Helpers.FtpPassword;
-                    connection.Port = Helpers.FtpPort;
-                    connection.SslMode = ftpsSslMode;
-                    connection.UseFTPS = true;
-                    connection.CertificateHashStringSHA1 = "nope";
-
-                    // Test and assert
-                    var ex = Assert.Throws<AggregateException>(() =>
-                    {
-                        var result = FTP.UploadFiles(source, destination, connection, new Options(), new Info(),
-                            new CancellationToken());
-                    });
-
-                    Assert.AreEqual(1, ex.InnerExceptions.Count);
-                    Assert.AreEqual(typeof(AuthenticationException), ex.InnerExceptions[0].GetType());
-                }
-            }
+            Assert.AreEqual(1, ex.InnerExceptions.Count);
+            Assert.AreEqual(typeof(AuthenticationException), ex.InnerExceptions[0].GetType());
         }
     }
 }
